+++ Sources/BraintreeDataCollector/BTDataCollector.m
@@ -101,7 +101,7 @@ NSString * const BTDataCollectorKountErrorDomain = @"com.braintreepayments.BTDat
         [self onCollectorStart];
 
         if (configuration.isKountEnabled && includeCard) {
-            BTDataCollectorEnvironment btEnvironment = [self environmentFromString:[configuration.json[@"environment"] asString]];
+            BTDataCollectorEnvironment btEnvironment = [self environmentFromString:configuration.environment];
             [self setCollectorEnvironment:[self collectorEnvironment:btEnvironment]];
 
             NSString *merchantID = self.fraudMerchantID ?: [configuration kountMerchantID];