+++ Sources/BraintreeVenmo/BTVenmoDriver.m
@@ -91,18 +91,15 @@ static BTVenmoDriver *appSwitchedDriver;
 
 #pragma mark - Tokenization
 
-- (void)authorizeAccountWithCompletion:(void (^)(BTVenmoAccountNonce *venmoAccount, NSError *error))completionBlock {
-    [self authorizeAccountAndVault:NO completion:completionBlock];
-}
+- (void)tokenizeVenmoAccountWithVenmoRequest:(BTVenmoRequest *)venmoRequest completion:(void (^)(BTVenmoAccountNonce * _Nullable venmoAccount, NSError * _Nullable error))completionBlock {
+    if (!venmoRequest) {
+        NSError *error = [NSError errorWithDomain:BTVenmoDriverErrorDomain
+                                             code:BTVenmoDriverErrorTypeIntegration
+                                         userInfo:@{NSLocalizedDescriptionKey: @"BTVenmoDriver failed because BTVenmoRequest is nil."}];
+        completionBlock(nil, error);
+        return;
+    }
 
-- (void)authorizeAccountAndVault:(BOOL)vault completion:(void (^)(BTVenmoAccountNonce *venmoAccount, NSError *error))completionBlock {
-    [self authorizeAccountWithProfileID:nil vault:vault completion:completionBlock];
-}
-
-- (void)authorizeAccountWithProfileID:(NSString *)profileID
-                                vault:(BOOL)vault
-                           completion:(void (^)(BTVenmoAccountNonce *venmoAccount, NSError *error))completionBlock
-{
     if (!self.apiClient) {
         NSError *error = [NSError errorWithDomain:BTVenmoDriverErrorDomain
                                              code:BTVenmoDriverErrorTypeIntegration
@@ -158,7 +155,7 @@ static BTVenmoDriver *appSwitchedDriver;
         [self informDelegateAppContextWillSwitch];
 
         [self.application openURL:appSwitchURL options:[NSDictionary dictionary] completionHandler:^(BOOL success) {
-            [self invokedOpenURLSuccessfully:success shouldVault:vault completion:completionBlock];
+            [self invokedOpenURLSuccessfully:success shouldVault:venmoRequest.vault completion:completionBlock];
         }];
     }];
 }