+++ CHANGELOG.md
@@ -27,8 +27,11 @@
   * Rename methods to use `ID` instead of `Id`:
     * `BTLocalPaymentRequest.localPaymentStarted(request:paymentID:start:)`
     * `BTVenmoDriver.authorizeAccount(profileID:vault:completion:)`
+  * Replace `BTVenmoDriver.authorizeAccount` methods with `BTVenmoDriver.tokenizeVenmoAccount`
 * Fix memory leak in `BTPayPalDriver`
 * Add `offerPayLater` to `BTPayPalRequest`
+* Add `environment` to `BTConfiguration`
+* Add `BTVenmoRequest`
 
 ## 5.0.0-beta2 (2021-01-20)
 * Add SPM support for `BraintreeDataCollector` and `BraintreeThreeDSecure`